quell warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@19005 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/FontManager.cxx
@@ -273,7 +273,7 @@ int FontManager::load(const char* fileName)
   }
 
   /* not found, add it */
-  id = fontFaces.size();
+  id = (int)fontFaces.size();
   fontFaces.push_back(face);
   name2id[face.name()] = id;
   file2id[fileName] = id;
